% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agro_2022}
\alias{agro_2022}
\title{Encuesta Nacional de Ingresos y Gastos de los Hogares 2022}
\format{
\subsection{\code{agro_2022}}{

A data frame with 1 rows and 3 columns:
\describe{
\item{Encuesta}{Name of the survey or data project.}
\item{Serie}{Series under which data is published.}
\item{Año}{Year span of data covered.}
...
}
}
}
\source{
\url{https://www.inegi.org.mx/programas/enigh/nc/2022/}
}
\usage{
agro_2022
}
\description{
A description
}
\keyword{datasets}
