% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2020.R, R/doc_2022.R
\docType{data}
\name{agroconsumo2020}
\alias{agroconsumo2020}
\alias{agroconsumo2022}
\title{Destino, cantidad y valor de los productos del negocio del hogar.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59985 rows and 11 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57456 rows and 11 columns.
}
\usage{
agroconsumo2020

agroconsumo2022
}
\description{
Permite identificar el destino, la cantidad y el valor de los productos del
negocio.
}
\details{
Identificar el destino, la cantidad y el valor estimado de los productos
agrícolas del negocio del hogar. Esta información se capta en el CUESTIONARIO
PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}, \code{id_trabajo}, \code{tipoact}
\code{numprod}, \code{codigo}, \code{cosecha}
\item \strong{Llave primaria}: \code{destino}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{agroproductos2022}
}
\keyword{datasets}
