% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2022.R
\docType{data}
\name{agrogasto_2022}
\alias{agrogasto_2022}
\title{Gasto que realiza el negocio del Hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60180 rows and 7 columns.
}
\usage{
agrogasto_2022
}
\description{
Permite identificar el tipo de gasto que realiza el negocio.
}
\details{
Gastos que realizan los negocios agrícolas del hogar. Esta información se
capta en el CUESTIONARIO PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}, \code{id_trabajo}, \code{tipoact}
\item \strong{Llave primaria}: \code{clave}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{agro_2022}
}
\keyword{datasets}
