% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2022.R
\docType{data}
\name{agroproductos_2022}
\alias{agroproductos_2022}
\title{Productos de los negocios agrícolas del hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69212 rows and 25 columns.
}
\usage{
agroproductos_2022
}
\description{
Se encuentran las características de los productos de los negocios agrícolas
de los hogares.
}
\details{
Características de los productos de los negocios agrícolas, forestales y de
tala, de cría y explotación y de pesca y caza de los hogares. Esta
información se capta en el CUESTIONARIO PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}, \code{id_trabajo}, \code{tipoact}
\item \strong{Llave primaria}: \code{numprod}, \code{codigo}, \code{cosecha}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{agro_2022}, \link{agroconsumo_2022}
}
\keyword{datasets}
