% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_survey.R
\name{as_survey}
\alias{as_survey}
\title{Survey sample ENIGH analysis}
\usage{
as_survey(data)
}
\arguments{
\item{data}{A data set from the ENIGH survey. It is important that the data
set includes the following variables: \code{est_dis} and \code{factor}.  \code{est_dis} is
the strata variable and \code{factor} is the weight variable.}
}
\value{
An object of class \code{survey.design}.
}
\description{
Turn an ENIGH data set into a survey design. Passes the data set to \code{\link[survey:svydesign]{survey::svydesign()}}.
Useful fore creating analyses with \code{\link[gtsummary:tbl_svysummary]{gtsummary::tbl_svysummary()}}.
}
\examples{
poblacion2022 |>
as_survey() |>
gtsummary::tbl_svysummary(include = c(nivelaprob))
}
