% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2018.R, R/doc_2020.R, R/doc_2022.R
\docType{data}
\name{concentradohogar_2018}
\alias{concentradohogar_2018}
\alias{concentradohogar_2020}
\alias{concentradohogar_2022}
\title{#' @rdname agroconsumo_2022
"agroconsumo_2018"}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74647 rows and 126 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 89006 rows and 126 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90102 rows and 126 columns.
}
\usage{
concentradohogar_2018

concentradohogar_2020

concentradohogar_2022
}
\description{
#' @rdname agrogasto_2022
"agrogasto_2018"
#' @rdname agroproductos_2022
"agroproductos_2018"

variables construidas a partir de las otras tablas de la base de datos.
Registra el resumen concentrado por hogar, de ingresos y gastos en toda
modalidad posible. Además todos los ingresos y gastos registrados en esta
tabla son trimestrales.
}
\details{
Esta tabla contiene las variables construidas, a partir de las tablas de la
base de datos, que permiten agrupar la información con la cual se obtienen
los principales tabulados que el INEGI construye.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares_2022}
}
\keyword{datasets}
