% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{enigh_documentation_2016}
\alias{enigh_documentation_2016}
\alias{enigh_documentation_2018}
\alias{enigh_documentation_2020}
\alias{enigh_documentation_2022}
\title{Documentación de la ENIGH}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 4 columns.
}
\usage{
enigh_documentation_2016

enigh_documentation_2018

enigh_documentation_2020

enigh_documentation_2022
}
\description{
Contiene una breve descripción de los distintos conjuntos de datos incluidos
en la ENIGH 2022.
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\keyword{datasets}
