% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2020.R, R/doc_2022.R
\docType{data}
\name{erogaciones_2020}
\alias{erogaciones_2020}
\alias{erogaciones_2022}
\title{Erogaciones de capital por hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67681 rows and 16 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64779 rows and 16 columns.
}
\usage{
erogaciones_2020

erogaciones_2022
}
\description{
Permite identificar las erogaciones financieras y de capital realizadas por
el hogar.
}
\details{
Erogaciones financieras y de capital captadas en el apartado 3.5 de la
sección III del CUESTIONARIO DE GASTOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}
\item \strong{Llave primaria}: \code{clave}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares_2022}
}
\keyword{datasets}
