% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{gastospersona2016}
\alias{gastospersona2016}
\alias{gastospersona2018}
\alias{gastospersona2020}
\alias{gastospersona2022}
\title{Gastos realizados en el hogar a nivel integrante}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 377013 rows and 20 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 398247 rows and 20 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 302694 rows and 20 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 402557 rows and 24 columns.
}
\usage{
gastospersona2016

gastospersona2018

gastospersona2020

gastospersona2022
}
\description{
Permite identificar los gastos realizados por cada integrante del hogar en
educación, transporte público y remuneraciones en especie.
}
\details{
Contiene los gastos monetarios y no monetarios asociados a los integrantes
del hogar en el periodo de referencia y que fueron captados en el
CUESTIONARIO DE GASTOS DEL HOGAR y en el CUADERNILLO DE GASTOS DIARIOS.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}
\item \strong{Llave primaria}: \code{clave}, \code{tipo_gasto}, \code{mes_dia},  \code{frec_rem}, \code{inst},
\code{forma_pag1}, \code{forma_pag2}, \code{forma_pag3}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{poblacion2022}
}
\keyword{datasets}
