% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2022.R
\docType{data}
\name{gastotarjetas_2022}
\alias{gastotarjetas_2022}
\title{Gastos realizados por el hogar y que fueron cubiertos mediante alguna
tarjeta de crédito bancaria y/o comercial}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13232 rows and 6 columns.
}
\usage{
gastotarjetas_2022
}
\description{
Permite identificar en qué medida los hogares se auxilian del financiamiento
que les ofrece la tarjeta de crédito bancaria o comercial para cubrir algunas
necesidades básicas.
}
\details{
Gastos realizados por el hogar y que fueron cubiertos mediante alguna tarjeta
de crédito bancaria y/o comercial. Estos fueron captados al final de cada
apartado del CUESTIONARIO DE GASTOS DEL HOGAR y en la sección 2 del
CUADERNILLO DE GASTOS DIARIOS.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}
\item \strong{Llave primaria}: \code{clave}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares_2022}
}
\keyword{datasets}
