% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2018.R, R/doc_2020.R, R/doc_2022.R
\docType{data}
\name{hogares_2018}
\alias{hogares_2018}
\alias{hogares_2020}
\alias{hogares_2022}
\title{Características de los hogares que habitan los integrantes del hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74647 rows and 137 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 89006 rows and 137 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90102 rows and 141 columns.
}
\usage{
hogares_2018

hogares_2020

hogares_2022
}
\description{
Se integra con información sobre el acceso a la alimentación de los
integrantes del hogar, el equipamiento del hogar, el tiempo de traslado al
hospital, los hábitos de consumo del hogar, las estimaciones de gasto en
alimentos y transporte, la información sobre los huéspedes y los
trabajadores domésticos. Esta se capta en las secciones II, IV, V y VI
del CUESTIONARIO DE HOGARES Y VIVIENDA y en el CUADERNILLO DE GASTOS DIARIOS.
}
\details{
Para la ENIGH 2012 y 2014 se separó la información referente a las viviendas
(antes contenida en la tabla HOGARES) en la tabla VIVIENDAS, de tal manera
que ahora la tabla HOGARES solo contiene características relativas al hogar.
Esta modificación corresponde a la incorporación de un factor de expansión
exclusivo para la vivienda dentro del diseño muestral. También se reestructuraron
las tablas correspondientes al gasto para respetar el nivel al que está asociada
la información; los gastos asociados al hogar, tanto monetarios como no monetarios,
se integraron en la tabla GASTOSHOGAR y los correspondientes a las personas se
integraron en la tabla GASTOSPERSONA. Esta modificación facilita la explotación
de la información referente al gasto, al encontrarse esta únicamente en dos tablas.
\itemize{
\item \strong{Llave primaria}: \code{foliohog}
\item \strong{Llave foránea}: \code{folioviv}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{concentradohogar_2022}, \link{poblacion_2022}, \link{gastotarjetas_2022},
\link{gastoshogar_2022}, \link{erogaciones_2022}
}
\keyword{datasets}
