% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{ingresos_2016}
\alias{ingresos_2016}
\alias{ingresos_2018}
\alias{ingresos_2020}
\alias{ingresos_2022}
\title{Ingresos y percepciones financieras y de capital de los integrantes
del hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 334337 rows and 17 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 348487 rows and 17 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 394912 rows and 17 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 397182 rows and 21 columns.
}
\usage{
ingresos_2016

ingresos_2018

ingresos_2020

ingresos_2022
}
\description{
permite identificar los ingresos y percepciones financieras y de capital de
cada uno de los integrantes del hogar, por diversos conceptos.
}
\details{
Ingresos y percepciones de capital de cada uno de los integrantes del hogar,
captados en los apartados referentes al ingreso en el CUESTIONARIO PARA
PERSONAS DE 12 O MÁS AÑOS y en el CUESTIONARIO PARA PERSONAS MENORES DE 12
AÑOS, así como los calculados a partir de la información del CUESTIONARIO
PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}
\item \strong{Llave primaria}: \code{clave}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares_2022}, \link{ingresos_jcf_2022}
}
\keyword{datasets}
