% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2022.R
\docType{data}
\name{ingresos_jcf_2022}
\alias{ingresos_jcf_2022}
\title{Ingresos del programa Jóvenes Construyendo el Futuro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 468 rows and 18 columns.
}
\usage{
ingresos_jcf_2022
}
\description{
Permite identificar los ingresos del programa Jóvenes Construyendo el Futuro.
}
\details{
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}
\item \strong{Llave primaria}: \code{clave}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{ingresos_2022}
}
\keyword{datasets}
