% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{noagro2016}
\alias{noagro2016}
\alias{noagro2018}
\alias{noagro2020}
\alias{noagro2022}
\title{#' @rdname ingresos_jcf2022
"ingresos_jcf2016"}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17969 rows and 105 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18984 rows and 105 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24666 rows and 115 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23847 rows and 115 columns.
}
\usage{
noagro2016

noagro2018

noagro2020

noagro2022
}
\description{
muestra a los trabajadores independientes mayores de 12 años, que tienen
negocios en el hogar dedicados a las actividades industriales, comerciales y
de servicios.
}
\details{
Ingresos y gastos de los negocios del hogar dedicados a las actividades
industriales, comerciales y de servicios, y sus características propias.
Esta información se capta en el CUESTIONARIO PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}, \code{id_trabajo}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogareserences
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{trabajos2022}, \link{noagroimportes2022}
}
\keyword{datasets}
