% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2020.R, R/doc_2022.R
\docType{data}
\name{noagroimportes_2020}
\alias{noagroimportes_2020}
\alias{noagroimportes_2022}
\title{Importe de los productos de los negocios no agrícolas del hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161874 rows and 17 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 155839 rows and 17 columns.
}
\usage{
noagroimportes_2020

noagroimportes_2022
}
\description{
Contiene los importes del negocio
}
\details{
Gastos que realizan los negocios no agrícolas de los hogares. Esta
información se capta en el CUESTIONARIO PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}, \code{id_trabajo}
\item \strong{Llave primaria}: \code{clave}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{noagro_2022}
}
\keyword{datasets}
