% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{poblacion2016}
\alias{poblacion2016}
\alias{poblacion2018}
\alias{poblacion2020}
\alias{poblacion2022}
\title{#' @rdname noagroimportes2022
"noagroimportes2016"}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 257805 rows and 178 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 269206 rows and 178 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 315743 rows and 184 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 309684 rows and 188 columns.
}
\usage{
poblacion2016

poblacion2018

poblacion2020

poblacion2022
}
\description{
Identifica las características sociodemográficas de los integrantes del
hogar; y el acceso a las instituciones de salud que tienen los integrantes
del hogar.
}
\details{
Características sociodemográficas y ocupacionales de los integrantes del
hogar. Esta información se capta en la sección III del CUESTIONARIO DE
HOGARES Y VIVIENDA, en los apartados 1.1 y 6.2 y secciones VIII, IX y X
del CUESTIONARIO PARA PERSONAS DE 12 O MÁS AÑOS y en la
sección II del CUESTIONARIO PARA PERSONAS MENORES DE 12 AÑOS.
\itemize{
\item \strong{Llave primaria}: \code{numren}
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares2022}, \link{ingresos2022}, \link{trabajos2022},
\link{gastospersona2022}
}
\keyword{datasets}
