% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{trabajos2016}
\alias{trabajos2016}
\alias{trabajos2018}
\alias{trabajos2020}
\alias{trabajos2022}
\title{Condición de actividad de los integrantes del hogar de 12 o más años}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 132095 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 139933 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 164876 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 165006 rows and 60 columns.
}
\usage{
trabajos2016

trabajos2018

trabajos2020

trabajos2022
}
\description{
Muestra la condición de actividad de los integrantes del hogar de 12 o más
años y algunas características ocupacionales durante el periodo de referencia.
}
\details{
Condición de actividad de los integrantes del hogar de 12 o más años. Se
integra con información captada en el CUESTIONARIO PARA PERSONAS DE 12 O MÁS
AÑOS.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}
\item \strong{Llave primaria}: \code{id_trabajo}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{poblacion2022}, \link{agro2022}, \link{noagro2022}
}
\keyword{datasets}
