% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2022.R
\docType{data}
\name{trabajos_2022}
\alias{trabajos_2022}
\title{Condición de actividad de los integrantes del hogar de 12 o más años}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 165006 rows and 60 columns.
}
\usage{
trabajos_2022
}
\description{
Muestra la condición de actividad de los integrantes del hogar de 12 o más
años y algunas características ocupacionales durante el periodo de referencia.
}
\details{
Condición de actividad de los integrantes del hogar de 12 o más años. Se
integra con información captada en el CUESTIONARIO PARA PERSONAS DE 12 O MÁS
AÑOS.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}
\item \strong{Llave primaria}: \code{id_trabajo}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{poblacion_2022}, \link{agro_2022}, \link{noagro_2022}
}
\keyword{datasets}
