% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{viviendas2016}
\alias{viviendas2016}
\alias{viviendas2018}
\alias{viviendas2020}
\alias{viviendas2022}
\title{Características de las viviendas que habitan los integrantes del hogar}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69169 rows and 64 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 73405 rows and 64 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 87754 rows and 64 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 88823 rows and 64 columns.
}
\usage{
viviendas2016

viviendas2018

viviendas2020

viviendas2022
}
\description{
Contiene las características de las viviendas que se captaron en la sección I del
CUESTIONARIO DE HOGARES Y VIVIENDA, el recuento de los residentes y los hogares
de la vivienda, la ubicación y el diseño muestral correspondiente a esta,
así como el factor de expansión.
}
\details{
Para la ENIGH 2012 y 2014 se separó la información referente a las viviendas
(antes contenida en la tabla HOGARES) en la tabla VIVIENDAS, de tal manera
que ahora la tabla HOGARES solo contiene características relativas al hogar.
Esta modificación corresponde a la incorporación de un factor de expansión
exclusivo para la vivienda dentro del diseño muestral. También se reestructuraron
las tablas correspondientes al gasto para respetar el nivel al que está asociada
la información; los gastos asociados al hogar, tanto monetarios como no monetarios,
se integraron en la tabla GASTOSHOGAR y los correspondientes a las personas se
integraron en la tabla GASTOSPERSONA. Esta modificación facilita la explotación
de la información referente al gasto, al encontrarse esta únicamente en dos tablas.
\itemize{
\item \strong{Llave primaria}: \code{folioviv}.
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares2022}
}
\keyword{datasets}
