% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{agro_2016}
\alias{agro_2016}
\alias{agro_2018}
\alias{agro_2020}
\alias{agro_2022}
\title{Negocios agropecuarios de los hogares}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13806 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14925 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18656 rows and 66 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17470 rows and 66 columns.
}
\usage{
agro_2016

agro_2018

agro_2020

agro_2022
}
\description{
Muestra la información de los trabajadores independientes, mayores de 12 años
que tienen en el hogar negocios dedicados a las actividades agrícolas,
forestales y de tala, además de actividades de cría, explotación y productos
derivados de la pesca y caza.
}
\details{
Ingresos y gastos de los negocios del hogar dedicados tanto a las actividades
agrícolas, forestales y de tala, como a actividades de cría, explotación y
productos derivados de la pesca y caza. Esta información se capta en el
CUESTIONARIO PARA NEGOCIOS DEL HOGAR.
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}, \code{numren}, \code{id_trabajo}
\item \strong{Llave primaria}: \code{tipoact}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{trabajos_2022}, \link{agroproductos_2022}, \link{agrogasto_2022}
}
\keyword{datasets}
