% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_2016.R, R/doc_2018.R, R/doc_2020.R,
%   R/doc_2022.R
\docType{data}
\name{gastoshogar_2016}
\alias{gastoshogar_2016}
\alias{gastoshogar_2018}
\alias{gastoshogar_2020}
\alias{gastoshogar_2022}
\title{Gastos en los hogares}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1048575 rows and 27 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1048575 rows and 27 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4948444 rows and 27 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5075174 rows and 31 columns.
}
\usage{
gastoshogar_2016

gastoshogar_2018

gastoshogar_2020

gastoshogar_2022
}
\description{
Permite identificar los gastos monetarios y no monetarios del hogar.
}
\details{
Contiene los gastos monetarios y no monetarios que realizó el hogar en el
periodo de referencia, así como la estimación del alquiler, que fueron
captados en el CUESTIONARIO DE GASTOS DEL HOGAR, en el CUADERNILLO DE GASTOS
DIARIOS y en el CUESTIONARIO DE HOGARESY VIVIENDA .
\itemize{
\item \strong{Llave foránea}: \code{folioviv}, \code{foliohog}
\item \strong{Llave primaria}: \code{clave}, \code{tipo_gasto}, \code{mes_dia}, \code{forma_pag1},
\code{forma_pag2}, \code{forma_pag3}, \code{lugar_comp}, \code{orga_inst},
\code{frecuencia}, \code{fecha_adqu}, \code{fecha_pago}
}
}
\references{
INEGI, 2023. Encuesta Nacional de Ingresos y Gastos de los Hogares
2022. ENIGH. Nueva Serie. \href{https://www.inegi.org.mx/app/biblioteca/ficha.html?upc=889463910626}{PDF}
}
\seealso{
\link{hogares_2022}
}
\keyword{datasets}
